refactor: localStorage 관련 함수 utils로 변경
