+++ src/util/index.ts
@@ -1,13 +1,11 @@
-const deepCopy = <T>(obj: T): T => {
-  let result: Partial<T> = {};
-  if (typeof obj === "object" && obj !== null) {
-    for (const key in obj) {
-      result[key] = deepCopy(obj[key]);
-    }
-  } else {
-    result = obj;
-  }
-  return result as T;
-};
+import { deepCopy } from "./deepCopy";
+import {
+  getResturantsFromLocalStorage,
+  setRestaurantsToLocalStorage,
+} from "./localStorage";
 
-export { deepCopy };
+export {
+  deepCopy,
+  getResturantsFromLocalStorage,
+  setRestaurantsToLocalStorage,
+};
