+++ src/util/localStorage.ts
@@ -1,0 +1,18 @@
+import { Restaurant } from "../types";
+
+const RESTAURANTS = "restaurants";
+
+export function getResturantsFromLocalStorage(): Restaurant[] {
+  if (!localStorage.getItem(RESTAURANTS)) {
+    localStorage.setItem(RESTAURANTS, JSON.stringify([]));
+  }
+
+  const restaurants = localStorage.getItem(RESTAURANTS) as string;
+
+  return JSON.parse(restaurants);
+}
+
+export function setRestaurantsToLocalStorage(newRestuarant: Restaurant) {
+  const newRestaurants = [...getResturantsFromLocalStorage(), newRestuarant];
+  localStorage.setItem(RESTAURANTS, JSON.stringify(newRestaurants));
+}
